# api_spec/openapi.yaml
openapi: "3.0.0"
servers:
  - url: http://localhost:8080/
info:
  title: Recipe Agent backend API service
  version: 1.0.0
paths:
...
  /ingredient:
    get:
      summary: Find ingredient(s)
      description: Provide attributes to list ingredients from the database. If no attribute is provided, it returns all.
      operationId: getIngredients
      parameters:
        - name: ingredient_id
          in: query
          required: false
          schema:
            type: integer
            example: 7
        - name: name
          in: query
          required: false
          schema:
            type: string
            example: carrot
        - name: ingredient_type
          in: query
          required: false
          schema:
            type: string
            example: carrot
            enum:
              - meat
              - dairy
              - vegetable
              - fruit
              - bakery
              - nuts
              - beverages
              - fats
      responses:
        '200':
          description: The query hits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingredient'
        '404':
          description: Not found
    post:
      summary: Add new ingredient
      description: Add a new ingredient into the database
      operationId: addIngredient
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ingredient'
        required: true
      responses:
        '200':
          description: Successful addition
          content:
            text/plain:
              schema:
                type: string
              example: 'Successfully added new ingredient!'
        '400':
          description: Invalid input
  /ingredient/{ingredient_id}:
    get:
      summary: Find ingredient by the ID
      description: THe ID is coming from the database
      operationId: getIngredientById
      parameters:
        - name: ingredient_id
          in: path
          required: true
          schema:
            type: integer
            example: 7
      responses:
        '200':
          description: The query hits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingredient'
        '404':
          description: Not found
    put:
      summary: Update ingredient
      description: Update the ingredient's attributes which belongs to the ID
      operationId: updateIngredientById
      parameters:
        - name: ingredient_id
          in: path
          required: true
          schema:
            type: integer
            example: 7
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ingredient'
        required: true
      responses:
        '200':
          description: The query hits
          content:
            text/plain:
              schema:
                type: string
              example: 'Successful modification of the ingredient!'
        '400':
          description: Invalid input
        '404':
          description: Not found
    delete:
      summary: Remove ingredient
      description: Remove the specified ingredient from the database
      operationId: deleteIngredientById
      parameters:
        - name: ingredient_id
          in: path
          required: true
          schema:
            type: integer
            example: 7
      responses:
        '200':
          description: OK
        '404':
          description: Not found
components:
  schemas:
    Ingredient:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 77
        name:
          type: string
          example: banana
        ingredient_type:
          type: string
          description: The type of the recipe ingredient
          example: meat
          enum:
            - meat
            - dairy
            - vegetable
            - fruit
            - bakery
            - nuts
            - beverages
            - fats
...